# target/classes/swagger/adexchange.yaml
...
swagger: '2.0'
info:
  title: TechM Arena Ad Exchange API
  description: API specification for TechM Arena Ad Exchange
  version: "1.0.0"
host: api.test.com
schemes:
  - http
  - https
basePath: /v1
produces:
  - application/json
paths:
...
  /role:
    get:
      description: |
        List of Role
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      tags:
        - User
      responses:
        200:
          description: List of Role
          schema:
            $ref: '#/definitions/RoleList'          
    post:
      description: Creates a new Role.
      operationId: createRole
      parameters:
        - name: role
          in: body
          description: Role to be created
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '200':
          description: created Role
          schema:
            $ref: '#/definitions/Role'
    put:
      description: Updates the Role.
      operationId: updateRole
      parameters:
        - name: role
          in: body
          description: Role to be updated
          required: true
          schema:
            $ref: '#/definitions/Role'
      responses:
        '200':
          description: updated Role
          schema:
            $ref: '#/definitions/Role'
  /role/{id}:
    get:
      description: Returns Role for the given id
      operationId: getRole
      parameters:
        - name: id
          in: path
          description: ID of Role
          required: true
          type: integer
      responses:
        '200':
          description: Role response
          schema:
            $ref: '#/definitions/Role'
    delete:
      description: deletes a single Role based on the ID supplied
      operationId: deleteRole
      parameters:
        - name: id
          in: path
          description: ID of Role to delete
          required: true
          type: integer
      responses:
        '200':
          description: Role deleted
  /role/search:
    get:
      description: Search Role
      operationId: searchRole
      parameters:
        - name: roleId
          in: query
          description: roleId
          required: false
          type: string
        - name: roleName
          in: query
          description: roleName
          required: false
          type: integer
        - name: roleGroup
          in: query
          description: roleGroup
          required: false
          type: integer
        - name: createdTimeStart
          in: query
          description: createdTimeStart
          required: false
          type: string
        - name: createdTimeEnd
          in: query
          description: createdTimeEnd
          required: false
          type: string
        - name: updatedTimeStart
          in: query
          description: updatedTimeStart
          required: false
          type: string
        - name: updatedTimeEnd
          in: query
          description: updatedTimeEnd
          required: false
          type: string
        - name: createdByUser
          in: query
          description: createdByUser
          required: false
          type: string
        - name: updatedByUser
          in: query
          description: updatedByUser
          required: false
          type: string
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of Role
          schema:
            $ref: '#/definitions/RoleList'
  /role/search/keyword:
    get:
      description: Search Role by keyword
      operationId: searchRoleByKeyword
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of Role
          schema:
            $ref: '#/definitions/RoleList'            
...
definitions:
...
  Role:
    type: object
    properties:
      id:
        type: integer
        description: Id
      roleId:
        type: string
      roleName:
        type: integer
      roleGroup:
        type: integer
      createdTime:
        type: string
        format: date
      updatedTime:
        type: string
        format: date
      createdByUser:
        type: string
      updatedByUser:
        type: string
  RoleList:
    type: object
    properties:
      count:
        type: integer
        format: int64
        description: Total count of objects
      roles:
        type: array
        items:
          $ref: '#/definitions/Role'      
...